package: convert internal ZIP handling data-types to 64bit
Prepare for a ZIP64 implementation.
Audit all "Size" property fetches through Anys.
Audit all uses of nSize, nCompressedSize, nOffset through the code.
Add FIXME64: comments to all points requiring future work.

+++ odk/examples/java/Storage/TestHelper.java
@@ -428,7 +428,7 @@ public class TestHelper  {
             {
                 // get "MediaType" and "Size" properties and control there values
                 String sPropMediaType = AnyConverter.toString( xPropSet.getPropertyValue( "MediaType" ) );
-                int nPropSize = AnyConverter.toInt( xPropSet.getPropertyValue( "Size" ) );
+                long nPropSize = AnyConverter.toLong( xPropSet.getPropertyValue( "Size" ) );
 
                 bOk = true;
                 if ( !sPropMediaType.equals( sMediaType ) )
